pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hiring-manager-tools"
version = "0.1.0"
description = "Spec files for AI-native people management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.49.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "mcp>=1.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]

[project.scripts]
spec = "hiring_manager_tools.cli:main"
